Optimise Start of LocationController to after Network is connected
GNSS generates a lot of UART data that trriggers many interrupts that slow the logon to wifi and MQTT connect to the Cloud. Starting the LocationController after connecting to Meadow Cloud makes overall startup to valiad logging of data faster and reduces number of attempts needed to sign on to Meadow Cloud. Added Property to store the last measured location.